refactor!: relocate schema introspection and validation into the Schema namespace
Schema introspection and validation lived under Services\ but have nothing to do with the action layer. SchemaIntrospector and its ColumnDefinition helper move to Schema\Introspection; SchemaValidator, SchemaValidationError and the validation rules move to Schema\Validation, so Services\ owns only the action layer. The contracts (SchemaValidationRule, SchemaIntrospectionProvider) stay in Contracts\ with their docblock references repointed. Every class body, method and signature is byte-for-byte unchanged - only namespaces, paths, imports and self-referential docblocks change - and all references across src and tests (container bindings, the validate-schemas command, exceptions, the schema compiler and OpenApi consumers) are updated. Tests mirror the new layout; the emptied Services\Introspection and Services\Validation directories are removed.

BREAKING CHANGE: SchemaIntrospector/SchemaValidator/SchemaValidationError/ColumnDefinition and the schema validation rules move from the Services namespace into Schema\Introspection and Schema\Validation.
