+++ src/Contracts/SchemaValidationRule.php
@@ -20,7 +20,7 @@ interface SchemaValidationRule
      * @param  string  $resourceClass
      * @param  string|null  $modelClass
      * @param  \SineMacula\ApiToolkit\Schema\CompiledSchema  $schema
-     * @return array<int, \SineMacula\ApiToolkit\Services\Validation\SchemaValidationError>
+     * @return array<int, \SineMacula\ApiToolkit\Schema\Validation\SchemaValidationError>
      */
     public function validate(string $resourceClass, ?string $modelClass, CompiledSchema $schema): array;
 }
