+++ src/Schema/Validation/Rules/ValidateComputedFields.php
@@ -2,19 +2,20 @@
 
 declare(strict_types = 1);
 
-namespace SineMacula\ApiToolkit\Services\Validation\Rules;
+namespace SineMacula\ApiToolkit\Schema\Validation\Rules;
 
 use SineMacula\ApiToolkit\Contracts\SchemaValidationRule;
 use SineMacula\ApiToolkit\Schema\CompiledSchema;
-use SineMacula\ApiToolkit\Services\Validation\SchemaValidationError;
+use SineMacula\ApiToolkit\Schema\Validation\SchemaValidationError;
 
 /**
- * Validate that relation resource class strings reference existing classes.
+ * Validate that computed field values are callable or reference valid methods
+ * on the resource class.
  *
  * @author      Ben Carey <bdmc@sinemacula.co.uk>
  * @copyright   2026 Sine Macula Limited.
  */
-final class ValidateRelationClasses implements SchemaValidationRule
+final class ValidateComputedFields implements SchemaValidationRule
 {
     /**
      * Validate the compiled schema for the given resource class.
@@ -22,7 +23,7 @@ final class ValidateRelationClasses implements SchemaValidationRule
      * @param  string  $resourceClass
      * @param  string|null  $modelClass
      * @param  \SineMacula\ApiToolkit\Schema\CompiledSchema  $schema
-     * @return array<int, \SineMacula\ApiToolkit\Services\Validation\SchemaValidationError>
+     * @return array<int, \SineMacula\ApiToolkit\Schema\Validation\SchemaValidationError>
      */
     #[\Override]
     public function validate(string $resourceClass, ?string $modelClass, CompiledSchema $schema): array
@@ -33,18 +34,22 @@ final class ValidateRelationClasses implements SchemaValidationRule
 
             $field = $schema->getField($key);
 
-            if ($field === null || $field->resource === null) {
+            if ($field === null || $field->compute === null) {
                 continue;
             }
 
-            if (class_exists($field->resource)) {
+            if ($field->compute instanceof \Closure || is_callable($field->compute)) {
+                continue;
+            }
+
+            if (is_string($field->compute) && method_exists($resourceClass, $field->compute)) {
                 continue;
             }
 
             $errors[] = new SchemaValidationError(
                 resourceClass: $resourceClass,
                 fieldKey: $key,
-                defect: sprintf('Relation resource class "%s" does not exist', $field->resource),
+                defect: 'Computed field value is not callable and does not reference an existing method on the resource class',
             );
         }
