+++ tests/Unit/OpenApi/Resolution/ColumnTypeMapperTest.php
@@ -8,8 +8,8 @@ use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use SineMacula\ApiToolkit\OpenApi\Resolution\ColumnTypeMapper;
+use SineMacula\ApiToolkit\Schema\Introspection\ColumnDefinition;
 use SineMacula\ApiToolkit\Schema\OpenApiFieldSchema;
-use SineMacula\ApiToolkit\Services\Introspection\ColumnDefinition;
 
 /**
  * Tests for the ColumnTypeMapper.
