+++ tests/Unit/Schema/Validation/Rules/ValidateComputedFieldsTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types = 1);
 
-namespace Tests\Unit\Services\Validation\Rules;
+namespace Tests\Unit\Schema\Validation\Rules;
 
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use SineMacula\ApiToolkit\Schema\CompiledFieldDefinition;
 use SineMacula\ApiToolkit\Schema\CompiledSchema;
-use SineMacula\ApiToolkit\Services\Validation\Rules\ValidateComputedFields;
+use SineMacula\ApiToolkit\Schema\Validation\Rules\ValidateComputedFields;
 use Tests\Fixtures\Resources\UserResource;
 
 /**
