+++ tests/Unit/Schema/Validation/Rules/ValidateGuardsTest.php
@@ -2,16 +2,17 @@
 
 declare(strict_types = 1);
 
-namespace Tests\Unit\Services\Validation\Rules;
+namespace Tests\Unit\Schema\Validation\Rules;
 
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
+use SineMacula\ApiToolkit\Schema\CompiledCountDefinition;
 use SineMacula\ApiToolkit\Schema\CompiledFieldDefinition;
 use SineMacula\ApiToolkit\Schema\CompiledSchema;
-use SineMacula\ApiToolkit\Services\Validation\Rules\ValidateTransformers;
+use SineMacula\ApiToolkit\Schema\Validation\Rules\ValidateGuards;
 
 /**
- * Tests for the ValidateTransformers validation rule.
+ * Tests for the ValidateGuards validation rule.
  *
  * @author      Ben Carey <bdmc@sinemacula.co.uk>
  * @copyright   2026 Sine Macula Limited.
@@ -20,15 +21,15 @@ use SineMacula\ApiToolkit\Services\Validation\Rules\ValidateTransformers;
  *
  * @internal
  */
-#[CoversClass(ValidateTransformers::class)]
-final class ValidateTransformersTest extends TestCase
+#[CoversClass(ValidateGuards::class)]
+final class ValidateGuardsTest extends TestCase
 {
     /**
-     * Test no errors for schema with callable transformers.
+     * Test no errors for schema with callable guards.
      *
      * @return void
      */
-    public function testNoErrorsForSchemaWithCallableTransformers(): void
+    public function testNoErrorsForSchemaWithCallableGuards(): void
     {
         $field = new CompiledFieldDefinition(
             accessor: 'name',
@@ -39,8 +40,8 @@ final class ValidateTransformersTest extends TestCase
             constraint: null,
             extras: [],
             needs: [],
-            guards: [],
-            transformers: [fn ($value, $model) => strtoupper($value)],
+            guards: [fn ($value, $model) => true],
+            transformers: [],
         );
 
         $schema = new CompiledSchema(
@@ -48,18 +49,18 @@ final class ValidateTransformersTest extends TestCase
             counts: [],
         );
 
-        $rule   = new ValidateTransformers;
+        $rule   = new ValidateGuards;
         $errors = $rule->validate('App\Http\Resources\UserResource', null, $schema);
 
         self::assertSame([], $errors);
     }
 
     /**
-     * Test reports non-callable transformer on field.
+     * Test reports non-callable guard on field.
      *
      * @return void
      */
-    public function testReportsNonCallableTransformerOnField(): void
+    public function testReportsNonCallableGuardOnField(): void
     {
         $field = new CompiledFieldDefinition(
             accessor: 'name',
@@ -70,8 +71,8 @@ final class ValidateTransformersTest extends TestCase
             constraint: null,
             extras: [],
             needs: [],
-            guards: [],
-            transformers: ['not_a_function'],
+            guards: ['not_a_function'],
+            transformers: [],
         );
 
         $schema = new CompiledSchema(
@@ -79,21 +80,50 @@ final class ValidateTransformersTest extends TestCase
             counts: [],
         );
 
-        $rule   = new ValidateTransformers;
+        $rule   = new ValidateGuards;
         $errors = $rule->validate('App\Http\Resources\UserResource', null, $schema);
 
         self::assertCount(1, $errors);
         self::assertSame('App\Http\Resources\UserResource', $errors[0]->resourceClass);
         self::assertSame('name', $errors[0]->fieldKey);
-        self::assertSame('Transformer at index 0 is not callable', $errors[0]->defect);
+        self::assertSame('Guard at index 0 is not callable', $errors[0]->defect);
     }
 
     /**
-     * Test reports multiple non-callable transformers.
+     * Test reports non-callable guard on count definition.
      *
      * @return void
      */
-    public function testReportsMultipleNonCallableTransformers(): void
+    public function testReportsNonCallableGuardOnCountDefinition(): void
+    {
+        $count = new CompiledCountDefinition(
+            presentKey: 'comments_count',
+            relation: 'comments',
+            constraint: null,
+            isDefault: false,
+            guards: ['not_a_function'],
+        );
+
+        $schema = new CompiledSchema(
+            fields: [],
+            counts: ['comments_count' => $count],
+        );
+
+        $rule   = new ValidateGuards;
+        $errors = $rule->validate('App\Http\Resources\PostResource', null, $schema);
+
+        self::assertCount(1, $errors);
+        self::assertSame('App\Http\Resources\PostResource', $errors[0]->resourceClass);
+        self::assertSame('comments_count', $errors[0]->fieldKey);
+        self::assertSame('Guard at index 0 is not callable', $errors[0]->defect);
+    }
+
+    /**
+     * Test reports multiple non-callable guards.
+     *
+     * @return void
+     */
+    public function testReportsMultipleNonCallableGuards(): void
     {
         $field = new CompiledFieldDefinition(
             accessor: 'email',
@@ -104,8 +134,8 @@ final class ValidateTransformersTest extends TestCase
             constraint: null,
             extras: [],
             needs: [],
-            guards: [],
-            transformers: ['invalid_one', 'invalid_two'],
+            guards: ['invalid_one', 'invalid_two'],
+            transformers: [],
         );
 
         $schema = new CompiledSchema(
@@ -113,20 +143,20 @@ final class ValidateTransformersTest extends TestCase
             counts: [],
         );
 
-        $rule   = new ValidateTransformers;
+        $rule   = new ValidateGuards;
         $errors = $rule->validate('App\Http\Resources\UserResource', null, $schema);
 
         self::assertCount(2, $errors);
-        self::assertSame('Transformer at index 0 is not callable', $errors[0]->defect);
-        self::assertSame('Transformer at index 1 is not callable', $errors[1]->defect);
+        self::assertSame('Guard at index 0 is not callable', $errors[0]->defect);
+        self::assertSame('Guard at index 1 is not callable', $errors[1]->defect);
     }
 
     /**
-     * Test no errors for schema with no transformers.
+     * Test no errors for schema with no guards.
      *
      * @return void
      */
-    public function testNoErrorsForSchemaWithNoTransformers(): void
+    public function testNoErrorsForSchemaWithNoGuards(): void
     {
         $field = new CompiledFieldDefinition(
             accessor: 'name',
@@ -146,7 +176,7 @@ final class ValidateTransformersTest extends TestCase
             counts: [],
         );
 
-        $rule   = new ValidateTransformers;
+        $rule   = new ValidateGuards;
         $errors = $rule->validate('App\Http\Resources\UserResource', null, $schema);
 
         self::assertSame([], $errors);
